feat: adapted existing library to parse and validate rrules
This is an elixir library scavanged from https://github.com/Flickswitch/exdateutils.
I've updated the underlying rust libraries, but much of the main library's (rrule) API has changed.

This library now has 3 main functions:
- parse rrule string into a struct of rrule properties
- parse rrule property structs into an rrule string
- validate rrule structs are valid

I am a rust novice so please don't be too hash on the code style,
but it seems to be working!